requirements.txt
Flask>=3.0.0
Flask-Bcrypt>=1.0.1
PyJWT>=2.8.0
